Fix role order column type from INTEGER to BIGINT
The order column was defined as INTEGER which has a max value of 2,147,483,647.
When existing roles had large order values (timestamps), creating new roles would
calculate Math.max(...order) + 1 which exceeded INTEGER range, causing:
"value '1761484391609' is out of range for type integer"

Changes:
- Changed roles.order from INTEGER to BIGINT
- BIGINT can store values up to 9,223,372,036,854,775,807
- Matches the same type used for created_at timestamps

Users will need to run ALTER TABLE to update existing databases.

+++ lib/db/schema.ts
@@ -19,7 +19,7 @@ CREATE TABLE IF NOT EXISTS roles (
   label TEXT NOT NULL,
   icon TEXT NOT NULL,
   slug TEXT NOT NULL,
-  "order" INTEGER NOT NULL,
+  "order" BIGINT NOT NULL,
   created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   UNIQUE(user_id, slug)
 );
